requirements.txt
aiodns==3.0.0
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
ccxt==2.8.21
ccxtpro==1.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.1
frozenlist==1.3.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
hyperframe==6.0.1
idna==3.4
kucoin-futures-python==1.0.6
kucoin-python==1.0.11
multidict==6.0.4
prettytable==3.6.0
psycopg2==2.9.5
pycares==4.3.0
pycparser==2.21
PySocks==1.7.1
pyTelegramBotAPI==4.10.0
python-dotenv==0.21.1
python-telegram-bot==20.1
redis==4.5.1
requests==2.28.2
rfc3986==1.5.0
sniffio==1.3.0
telebot==0.0.5
urllib3==1.26.14
wcwidth==0.2.6
websockets==10.4
yarl==1.8.2
